requirements.txt
Flask==2.0.3
flask_restfulflask_cors
wordcloud
transformers
googletrans
pprint

